Update README.md
About BarbSal_MS - The purpose of a salon appointment application is to provide a convenient and efficient way for customers to schedule appointments and manage their salon visits. The application serves as a digital platform that connects customers with salons, spas, or hairdressers, facilitating the booking process and enhancing the overall customer experience.